Implementando o começo do fluxo de pagamento parte 1

+++ src/main/java/br/com/zup/academy/ednelson/casadocodigo/pais/PaisRepository.java
@@ -2,6 +2,6 @@ package br.com.zup.academy.ednelson.casadocodigo.pais;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PaisRepository extends JpaRepository<Pais, Integer> {
+public interface PaisRepository extends JpaRepository<Pais, Long> {
 
 }
